+++ integration-tests/src/tests/client/features/in_memory_tries.rs
@@ -2,8 +2,8 @@ use near_async::messaging::CanSend;
 use near_async::time::{FakeClock, Utc};
 use near_chain::{Block, Provenance};
 use near_chain_configs::{Genesis, GenesisConfig, GenesisRecords};
+use near_chunks::shards_manager_actor::CHUNK_REQUEST_SWITCH_TO_FULL_FETCH;
 use near_chunks::test_loop::ShardsManagerResendChunkRequests;
-use near_chunks::CHUNK_REQUEST_SWITCH_TO_FULL_FETCH;
 use near_client::test_utils::TestEnv;
 use near_client::ProcessTxResponse;
 use near_o11y::testonly::init_test_logger;
